Convert announcements and comments API wrappers to makeApiCall
Behavior-preserving: same exported names, params/opts signatures,
return shapes, and functionName error context. Also converts both
files from Flow to TypeScript.
